Replace int indices with size_t type

+++ src/utils/file_utils.h
@@ -7,6 +7,6 @@ namespace file_utils
 {
     std::vector<std::string> listSubdirectories(std::string path);
     std::vector<std::string> listFiles(std::string path);
-    std::vector<std::string> getNRandomFilesFromSubdirectories(int n, std::vector<std::string> directories);
-    std::vector<std::string> getNRandomFiles(int n, std::string directory);
+    std::vector<std::string> getNRandomFilesFromSubdirectories(size_t n, std::vector<std::string> directories);
+    std::vector<std::string> getNRandomFiles(size_t n, std::string directory);
 }
